+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# NEXT
+# 3.0.0-beta.10
 - `Follower` mixin improvements.
 - Adds `startFollowPlayer` param in `CameraConfig`.
 - `Vision` improvements.